iot-simulator/requirements.txt
# azure-iot-location-monitoring\iot-simulator\requirements.txt  

azure-iot-device==2.11.0